Fix resolve field type on double field types (#29)
* Fix resolve field type on double field types

It is completely valid for the field type to be missing; some compilers (eg, js) will not include default enum values in the message, and in this case the default field type is double. When a double is sent currently, a missing field type error is returned. This error should never be returned. As is done elsewhere, we should use the getter instead of reading the field directly so that the default value is automatically included where needed.

* Add tests for missing type field